Additional status codes and bad arguments check

+++ lib/include/vector_ops.h
@@ -7,11 +7,11 @@
 extern "C" {
 #endif
 
-size_t sum(int a, int b);
+int32_t sum(int a, int b);
 
-size_t cum_sum(const int *vector_ptr, size_t size);
+int32_t cum_sum(const int *vector_ptr, size_t size);
 
-size_t sum_vectors(const int *vector_ptr_1, const int *vector_ptr_2, int *vector_dst_ptr, size_t size);
+int32_t sum_vectors(const int *vector_ptr_1, const int *vector_ptr_2, int *vector_dst_ptr, size_t size);
 
 #ifdef __cplusplus
 }
